@@ -3,93 +3,89 @@ declare(strict_types=1);
 
 namespace Trejjam\Configuration\DI;
 
-use Nette\Utils\Validators;
+use DateTimeImmutable;
+use Nette\DI\CompilerExtension;
+use Nette\PhpGenerator\Literal;
+use Nette\Schema\Expect;
+use Nette\Schema\Schema;
 use Trejjam;
 use Nextras\Migrations;
 
-final class ConfigurationExtension extends Trejjam\BaseExtension\DI\BaseExtension
+final class ConfigurationExtension extends CompilerExtension
 {
-	protected $default = [
-		'environment'  => Trejjam\Configuration\Environment::SITE_MODE_PUBLIC,
-		'isCli'        => PHP_SAPI === 'cli',
-		'fileVersion'  => [
-			'version'   => Trejjam\Configuration\FileVersion::UNSPECIFIED_VERSION,
-			'buildTime' => NULL,
-		],
-		'useMigration' => FALSE,
-		'migration'    => [
-			'dir'                 => NULL,
-			'withDummyData'       => FALSE,
-			'phpParams'           => [],
-			'disablePhpExtension' => FALSE,
-			'withTestData'        => FALSE,
-		],
-	];
+    public function getConfigSchema(): Schema
+    {
+        return Expect::structure([
+            'environment' => Expect::string()->default(Trejjam\Configuration\Environment::SITE_MODE_PUBLIC),
+            'isCli' => Expect::bool()->default(PHP_SAPI === 'cli'),
+            'fileVersion' => Expect::structure([
+                'version' => Expect::string()->default(Trejjam\Configuration\FileVersion::UNSPECIFIED_VERSION),
+                'buildTime' => Expect::string()->default(new DateTimeImmutable()),
+            ]),
+            'useMigration' => Expect::bool()->default(interface_exists(Migrations\IConfiguration::class)),
+            'migration' => Expect::structure([
+                'dir' => Expect::anyOf(
+                    Expect::string()->nullable()->default(null),
+                    Expect::type(Literal::class)->nullable()->default(null)
+                ),
+                'withDummyData' => Expect::bool()->default(false),
+                'phpParams' => Expect::array(),
+                'disablePhpExtension' => Expect::bool()->default(false),
+                'withTestData' => Expect::bool()->default(false),
+            ]),
+        ]);
+    }
 
-	protected $classesDefinition = [
-		'environment'     => Trejjam\Configuration\Environment::class,
-		'fileVersion'     => Trejjam\Configuration\FileVersion::class,
-		'systemDirectory' => Trejjam\Configuration\SystemDirectory::class,
-	];
+    public function beforeCompile()
+    {
+        $containerBuilder = $this->getContainerBuilder();
 
-	public function loadConfiguration(bool $validateConfig = TRUE) : void
-	{
-		$this->default['fileVersion']['buildTime'] = new \DateTimeImmutable;
-		$this->default['useMigration'] = interface_exists(Migrations\IConfiguration::class);
+        if ($this->config->useMigration) {
+            $migration = $containerBuilder->getDefinitionByType(Migrations\IConfiguration::class);
 
-		parent::loadConfiguration($validateConfig);
+            $migrationConfiguration = $this->config->migration;
 
-		Validators::assertField($this->config, 'isCli', 'bool');
-		Validators::assertField($this->config, 'useMigration', 'bool');
-		if ($this->config['useMigration']) {
-			Validators::assertField($this->config['migration'], 'dir', 'string|Nette\PhpGenerator\PhpLiteral');
-			Validators::assertField($this->config['migration'], 'phpParams', 'array');
-		}
+            $migration->setType(Trejjam\Configuration\Migration::class);
+            $migration->setFactory(
+                Trejjam\Configuration\Migration::class,
+                [
+                    'dir' => $migrationConfiguration['dir'],
+                    'withDummyData' => $migrationConfiguration['withDummyData'],
+                    'phpParams' => $migrationConfiguration['phpParams'],
+                    'disablePhpExtension' => $migrationConfiguration['disablePhpExtension'],
+                    'withTestData' => $migrationConfiguration['withTestData'],
+                ]
+            );
+        }
 
-		$containerBuilder = $this->getContainerBuilder();
-		$applicationParameters = $containerBuilder->parameters;
+        $containerBuilder->addDefinition($this->prefix('environment'))
+            ->setType(Trejjam\Configuration\Environment::class)
+            ->setArguments(
+                [
+                    'siteMode' => $this->config->environment,
+                    'isCli' => $this->config->isCli,
+                ]
+            );
 
-		$types = $this->getTypes();
+        $fileVersion = $this->config->fileVersion;
+        $containerBuilder->addDefinition($this->prefix('fileVersion'))
+            ->setType(Trejjam\Configuration\FileVersion::class)
+            ->setArguments(
+                [
+                    'version' => $fileVersion['version'],
+                    'buildTime' => $fileVersion['buildTime'],
+                ]
+            );
 
-		$types['environment']->setArguments(
-			[
-				'siteMode' => $this->config['environment'],
-				'isCli'    => $this->config['isCli'],
-			]
-		);
-		$types['fileVersion']->setArguments(
-			[
-				'version'   => $this->config['fileVersion']['version'],
-				'buildTime' => $this->config['fileVersion']['buildTime'],
-			]
-		);
-		$types['systemDirectory']->setArguments(
-			[
-				'tempDir' => $applicationParameters['tempDir'],
-				'appDir'  => $applicationParameters['appDir'],
-				'wwwDir'  => $applicationParameters['wwwDir'],
-			]
-		);
-	}
-
-	public function beforeCompile()
-	{
-		$containerBuilder = $this->getContainerBuilder();
-
-		if ($this->config['useMigration']) {
-			$migration = $containerBuilder->getDefinitionByType(Migrations\IConfiguration::class);
-
-			$migration->setType(Trejjam\Configuration\Migration::class);
-			$migration->setFactory(
-				Trejjam\Configuration\Migration::class,
-				[
-					'dir'                 => $this->config['migration']['dir'],
-					'withDummyData'       => $this->config['migration']['withDummyData'],
-					'phpParams'           => $this->config['migration']['phpParams'],
-					'disablePhpExtension' => $this->config['migration']['disablePhpExtension'],
-					'withTestData'        => $this->config['migration']['withTestData'],
-				]
-			);
-		}
-	}
+        $applicationParameters = $containerBuilder->parameters;
+        $containerBuilder->addDefinition($this->prefix('systemDirectory'))
+            ->setType(Trejjam\Configuration\SystemDirectory::class)
+            ->setArguments(
+                [
+                    'tempDir' => $applicationParameters['tempDir'],
+                    'appDir' => $applicationParameters['appDir'],
+                    'wwwDir' => $applicationParameters['wwwDir'],
+                ]
+            );
+    }
 }
