@@ -6,10 +6,10 @@ namespace Trejjam\Configuration\Helper\Migration;
 use Nextras\Migrations\Entities\File;
 use Nextras\Migrations\IExtensionHandler;
 
-class DummyHandler implements IExtensionHandler
+final class DummyHandler implements IExtensionHandler
 {
 	public function execute(File $file)
 	{
-		return NULL;
+		return null;
 	}
 }
